Replace short by int

+++ VarsViewer/VarsViewer/Var.cs
@@ -13,8 +13,8 @@ namespace VarsViewer
 
 		public int Index;
 		public string Name;
-		public short Value;
-		public short SaveState; //value set there when using SaveState button
+		public int Value;
+		public int SaveState; //value set there when using SaveState button
 		public int Time;	//time since last difference
 		public long MemoryAddress = -1;
 	}
